Restructuring of SVG's and useScrollTrack: - Have moved all drawable SVG's into their own component. - Have moved debounce from inside of useScrollTrack to the utilFunctions folder.
