+++ src/components/DrawableSVG/DrawableSVG.scss
@@ -1,0 +1,86 @@
+@use '../../styles/fonts' as *;
+@use '../../styles/variables' as *;
+@use '../../styles/mixins' as *;
+
+.svg__lock{
+  display: none;
+  position: absolute;
+  width: 25vmax;
+  top: -8vmin;
+  right: 25vmax;
+
+  @include tablet{
+    display: block;
+  }
+
+  @include desktop{
+    width: 40vmin;
+    right: 20vmax;
+  }
+
+  &.--active path{
+    stroke-dashoffset: 0;
+  }
+  
+  &-path{
+    stroke-dasharray: 1490;
+    stroke-dashoffset: 1490;
+    transition: stroke-dashoffset 2s ease-in-out;
+  }
+}
+
+.svg__question-mark{
+  display: none;
+  position: absolute;
+  width: 20vmax;
+  top: -8vmin;
+  left: 15vmax;
+
+  @include tablet{
+    display: block;
+  }
+
+  @include desktop{
+    width: 40vmin;
+    bottom: 20vmax;
+    left: 20vmax;
+  }
+
+  &.--active path{
+    stroke-dashoffset: 0;
+  }
+  
+  &-path{
+    stroke-dasharray: 956;
+    stroke-dashoffset: 956;
+    transition: stroke-dashoffset 2s ease-in-out;
+  }
+}
+
+.svg__phone{
+  display: none;
+  position: absolute;
+  width: 35vmin;
+  top: -15vmin;
+  right: 25vmin;
+
+  @include tablet{
+    display: block;
+  }
+
+  @include desktop{
+    width: 45vmin;
+    bottom: 10vmax;
+    left: 35vmax;
+  }
+
+  &.--active path{
+    stroke-dashoffset: 0;
+  }
+
+  &-path{
+    stroke-dasharray: 2255;
+    stroke-dashoffset: 2255;
+    transition: stroke-dashoffset 2s ease-in-out;
+  }
+}
